Remove ForceReceiveState from AtatCmd & remove old digester
