+++ .github/workflows/ci.yml
@@ -147,13 +147,7 @@ jobs:
         uses: actions-rs/cargo@v1
         with:
           command: test
-          args: --lib
-      # TODO: Change this to a single --all test, when the examples work
-      - name: Doctests
-        uses: actions-rs/cargo@v1
-        with:
-          command: test
-          args: --doc
+          args: --all
 
   grcov:
     name: Coverage
